requirements.txt
python-docx>=0.8.11
PyPDF2>=2.10.0
pdfplumber>=0.7.0
openai>=1.0.0
requests>=2.28.0
nltk>=3.7
jieba>=0.42.1
tqdm>=4.64.0
python-dotenv>=0.20.0
argparse>=1.4.0
colorama>=0.4.5
reportlab>=3.6